+++ site/admin/index.php
@@ -1,10 +1,10 @@
 <?php
   session_start();
   if(isset($_SESSION["User_ID"])){
-    //Login überprüfen
+    //Check login
     $logina = 0;
     require_once ('configuration.php');
-    $db_link = mysqli_connect (MYSQL_HOST,MYSQL_BENUTZER,MYSQL_KENNWORT,MYSQL_DATENBANK);
+    $db_link = mysqli_connect (MYSQL_HOST,MYSQL_USER,MYSQL_PASSWORD,MYSQL_DATABASE);
     $sql = "SELECT * FROM User Where Username = '".$_SESSION["User_Name"]."'";
     $db_erg = mysqli_query( $db_link, $sql );
     while ($zeile = mysqli_fetch_array( $db_erg, MYSQLI_ASSOC)){
@@ -14,13 +14,13 @@
     }
     if ($logina == 1){
       if(isset($_GET["URL"])&&file_exists("pages/".$_GET["URL"].".php")){
-        include "pages/".$_GET["URL"].".php";
+        include_once "pages/".$_GET["URL"].".php";
       }else{
-        include "pages/mainpage.php";
+        include_once "pages/mainpage.php";
       }
     }else{
       header("HTTP/1.1 403 Forbidden");
-      header('Location: '.getDomain().'/Errors/403.html');
+      header('Location: '.$USOC["DOMAIN"].'/Errors/403.html');
     }
   }else{
     echo <<<HTML